+++ modules/front-end/src/app/features/safe/experiments/types.ts
@@ -1,9 +1,3 @@
-import {
-  CustomEventSuccessCriteria,
-  CustomEventTrackOption,
-  EventType, ExperimentStatus,
-  IMetric
-} from "@features/safe/feature-flags/types/experimentations";
 import {IVariation} from "@shared/rules";
 
 export class MetricListFilter {
@@ -121,4 +115,52 @@ export interface IExptIterationResult {
 export interface IExptStatusCount {
   status: ExperimentStatus,
   count: number
+}
+
+export enum EventType {
+  Custom = 1,
+  PageView = 2,
+  Click = 3
+}
+
+export enum CustomEventTrackOption {
+  Undefined = 0,
+  Conversion = 1,
+  Numeric = 2
+}
+
+export enum CustomEventSuccessCriteria {
+  Undefined = 0,
+  Lower = 2, // lower than baseline
+  Higher = 1 // higher than baseline
+}
+
+export enum ExperimentStatus {
+  NotStarted = 'NotStarted',
+  Paused = 'Paused',
+  Recording = 'Recording'
+}
+
+export enum UrlMatchType {
+  Substring = 1
+}
+
+export interface ITargetUrl {
+  matchType: UrlMatchType,
+  url: string
+}
+
+export interface IMetric {
+  id?: string,
+  name: string,
+  envId: string,
+  description: string,
+  maintainerUserId: string,
+  eventName: string,
+  eventType: EventType,
+  customEventTrackOption?: CustomEventTrackOption,
+  customEventUnit?: string,
+  customEventSuccessCriteria?: CustomEventSuccessCriteria,
+  elementTargets?: string,
+  targetUrls?: ITargetUrl[]
 }
