+++ modules/front-end/src/app/features/safe/feature-flags/details/experimentation/experimentation.component.ts
@@ -2,7 +2,6 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { ExperimentService } from '@services/experiment.service';
-import { CustomEventTrackOption, EventType, ExperimentStatus } from '../../types/experimentations';
 import * as moment from 'moment';
 import {FeatureFlagService} from "@services/feature-flag.service";
 import {FeatureFlag, IFeatureFlag} from "@features/safe/feature-flags/types/details";
@@ -11,7 +10,10 @@ import {
   ExperimentListFilter,
   IExpt,
   IExptIteration,
-  IExptIterationResult
+  IExptIterationResult,
+  CustomEventTrackOption,
+  EventType,
+  ExperimentStatus
 } from "@features/safe/experiments/types";
 
 @Component({
